+++ components/archive/search/src/main/java/com/flipperdevices/archive/search/model/SearchNavigationConfig.kt
@@ -1,12 +1,9 @@
 package com.flipperdevices.archive.search.model
 
-import com.flipperdevices.archive.api.SelectKeyPathListener
 import kotlinx.serialization.Serializable
 
 @Serializable
 sealed class SearchNavigationConfig {
     @Serializable
-    data class Search(
-        val onItemSelected: SelectKeyPathListener?
-    ) : SearchNavigationConfig()
+    data object Search : SearchNavigationConfig()
 }