+++ components/bridge/connection/common/api/src/main/java/com/flipperdevices/bridge/connection/common/api/DeviceConnectionApi.kt
@@ -3,5 +3,9 @@ package com.flipperdevices.bridge.connection.common.api
 import kotlinx.coroutines.CoroutineScope
 
 interface DeviceConnectionApi<API : FConnectedDeviceApi, CONFIG : FDeviceConnectionConfig<API>> {
-    suspend fun connect(scope: CoroutineScope, config: CONFIG): Result<API>
+    suspend fun connect(
+        scope: CoroutineScope,
+        config: CONFIG,
+        listener: FTransportConnectionStatusListener
+    ): Result<API>
 }