+++ components/bridge/connection/sample/src/main/java/com/flipperdevices/bridge/connection/screens/models/ConnectionRootConfig.kt
@@ -9,4 +9,7 @@ sealed class ConnectionRootConfig {
 
     @Serializable
     data object NoPermission : ConnectionRootConfig()
+
+    @Serializable
+    data class Benchmark(val address: String) : ConnectionRootConfig()
 }