+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - [FIX] Update flipper is busy image
 - [Feature] Add not connected, empty and syncing states to emulation button on key screen
 - [Feature] Check app exist on apps catalog manifest loading
+- [Feature] First version of device orchestrator
 - [Refactor] Use https://github.com/IlyaGulya/anvil-utils to automatically generate and contribute assisted factories.
 - [Refactor] Add test for Progress Tracker logic